Remove dependencies pulled in when writing to output channels
