+++ packages/output/src/browser/output-contribution.ts
@@ -25,8 +25,9 @@ import { OutputWidget } from './output-widget';
 import { OutputContextMenu } from './output-context-menu';
 import { OutputUri } from '../common/output-uri';
 import { ClipboardService } from '@theia/core/lib/browser/clipboard-service';
-import { OutputChannelManager } from '../common/output-channel';
+import { OutputChannelManager, OutputChannel } from '../common/output-channel';
 import { OutputCommands } from './output-commands';
+import { QuickPickService, QuickPickItem } from '@theia/core/lib/common/quick-pick-service';
 
 @injectable()
 export class OutputContribution extends AbstractViewContribution<OutputWidget> implements OpenHandler {
@@ -42,6 +43,9 @@ export class OutputContribution extends AbstractViewContribution<OutputWidget> i
 
     @inject(OpenerService)
     protected readonly openerService: OpenerService;
+
+    @inject(QuickPickService)
+    protected readonly quickPickService: QuickPickService;
 
     readonly id: string = `${OutputWidget.ID}-opener`;
 
@@ -111,6 +115,105 @@ export class OutputContribution extends AbstractViewContribution<OutputWidget> i
                 }
             }
         });
+        registry.registerCommand(OutputCommands.APPEND, {
+            execute: ({ name, text }: { name: string, text: string }) => {
+                if (name && text) {
+                    this.outputChannelManager.getChannel(name).append(text);
+                }
+            }
+        });
+        registry.registerCommand(OutputCommands.APPEND_LINE, {
+            execute: ({ name, text }: { name: string, text: string }) => {
+                if (name && text) {
+                    this.outputChannelManager.getChannel(name).appendLine(text);
+                }
+            }
+        });
+        registry.registerCommand(OutputCommands.CLEAR, {
+            execute: ({ name }: { name: string }) => {
+                if (name) {
+                    this.outputChannelManager.getChannel(name).clear();
+                }
+            }
+        });
+        registry.registerCommand(OutputCommands.DISPOSE, {
+            execute: ({ name }: { name: string }) => {
+                if (name) {
+                    this.outputChannelManager.deleteChannel(name);
+                }
+            }
+        });
+        registry.registerCommand(OutputCommands.SHOW, {
+            execute: ({ name, options }: { name: string, options?: { preserveFocus?: boolean } }) => {
+                if (name) {
+                    const preserveFocus = options && options.preserveFocus || false;
+                    this.outputChannelManager.getChannel(name).show({ preserveFocus });
+                }
+            }
+        });
+        registry.registerCommand(OutputCommands.HIDE, {
+            execute: ({ name }: { name: string }) => {
+                if (name) {
+                    this.outputChannelManager.getChannel(name).hide();
+                }
+            }
+        });
+
+        registry.registerCommand(OutputCommands.CLEAR__QUICK_PICK, {
+            execute: async () => {
+                const channel = await this.pick({
+                    placeholder: 'Clear output channel.',
+                    channels: this.outputChannelManager.getChannels().slice()
+                });
+                if (channel) {
+                    channel.clear();
+                }
+            },
+            isEnabled: () => !!this.outputChannelManager.getChannels().length,
+            isVisible: () => !!this.outputChannelManager.getChannels().length
+        });
+        registry.registerCommand(OutputCommands.SHOW__QUICK_PICK, {
+            execute: async () => {
+                const channel = await this.pick({
+                    placeholder: 'Show output channel.',
+                    channels: this.outputChannelManager.getChannels().slice()
+                });
+                if (channel) {
+                    const { name } = channel;
+                    registry.executeCommand(OutputCommands.SHOW.id, { name, options: { preserveFocus: false } });
+                }
+            },
+            isEnabled: () => !!this.outputChannelManager.getChannels().length,
+            isVisible: () => !!this.outputChannelManager.getChannels().length
+        });
+        registry.registerCommand(OutputCommands.HIDE__QUICK_PICK, {
+            execute: async () => {
+                const channel = await this.pick({
+                    placeholder: 'Hide output channel.',
+                    channels: this.outputChannelManager.getVisibleChannels().slice()
+                });
+                if (channel) {
+                    const { name } = channel;
+                    registry.executeCommand(OutputCommands.HIDE.id, { name });
+                }
+            },
+            isEnabled: () => !!this.outputChannelManager.getVisibleChannels().length,
+            isVisible: () => !!this.outputChannelManager.getVisibleChannels().length
+        });
+        registry.registerCommand(OutputCommands.DISPOSE__QUICK_PICK, {
+            execute: async () => {
+                const channel = await this.pick({
+                    placeholder: 'Close output channel.',
+                    channels: this.outputChannelManager.getChannels().slice()
+                });
+                if (channel) {
+                    const { name } = channel;
+                    registry.executeCommand(OutputCommands.DISPOSE.id, { name });
+                }
+            },
+            isEnabled: () => !!this.outputChannelManager.getChannels().length,
+            isVisible: () => !!this.outputChannelManager.getChannels().length
+        });
     }
 
     registerMenus(registry: MenuModelRegistry): void {
@@ -150,4 +253,19 @@ export class OutputContribution extends AbstractViewContribution<OutputWidget> i
     ): boolean | false {
         return widget instanceof OutputWidget ? predicate(widget) : false;
     }
+
+    protected async pick({ channels, placeholder }: { channels: OutputChannel[], placeholder: string }): Promise<OutputChannel | undefined> {
+        const items: QuickPickItem<OutputChannel>[] = [];
+        for (let i = 0; i < channels.length; i++) {
+            const channel = channels[i];
+            if (i === 0) {
+                items.push({ label: channel.isVisible ? 'Output Channels' : 'Hidden Channels', type: 'separator' });
+            } else if (!channel.isVisible && channels[i - 1].isVisible) {
+                items.push({ label: 'Hidden Channels', type: 'separator' });
+            }
+            items.push({ label: channel.name, value: channel });
+        }
+        return this.quickPickService.show(items, { placeholder });
+    }
+
 }
